% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{get_acs_race_ethnicity}
\alias{get_acs_race_ethnicity}
\title{Access race and ethnicity data from the American Community Survey}
\usage{
get_acs_race_ethnicity(clean_variable_names = FALSE, ...)
}
\arguments{
\item{clean_variable_names}{Should variable names be cleaned (i.e. snake case)}

\item{...}{Other arguments passed to tidycensus::get_acs()}
}
\value{
A tibble with five variables: GEOID, NAME, variable, estimate, and moe
}
\description{
Access race and ethnicity data from the American Community Survey
}
