% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_dk}
\alias{theme_dk}
\title{David Keyes's custom ggplot theme}
\usage{
theme_dk(show_grid_lines = TRUE, show_axis_titles = TRUE)
}
\arguments{
\item{show_grid_lines}{Should grid lines be shown (TRUE/FALSE)}

\item{show_axis_titles}{Should axis titles be shown (TRUE/FALSE)}
}
\value{
A complete ggplot2 theme
}
\description{
David Keyes's custom ggplot theme
}
